requirements.txt
pybit==5.6.0
pandas==2.2.1
ta==0.11.0
python-dotenv==1.0.1
requests==2.31.0
numpy==1.26.4 